# crates/eidolons-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eidolons API",
    "description": "Privacy-transparent AI proxy API with inline verification metadata",
    "contact": {
      "name": "Mike Marcacci"
    },
    "license": {
      "name": "",
      "identifier": "UNLICENSED"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/account/balances": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account/balances — get credit balance breakdown (authenticated).",
        "operationId": "get_balances",
        "responses": {
          "200": {
            "description": "Balance breakdown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalancesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
...
    "/v1/account/ledger": {
      "get": {
        "tags": [
          "Linked"
        ],
        "summary": "GET /v1/account/ledger — get credit ledger entries (authenticated).",
        "operationId": "get_ledger",
        "parameters": [
          {
            "name": "reason",
            "in": "query",
            "description": "Comma-separated reason filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque pagination cursor from a previous response",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max entries to return (default 50, max 200)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ledger entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LedgerResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "basic": []
          }
        ]
      }
    },
...
    "/v1/chat/completions": {
      "post": {
        "tags": [
          "Unlinked"
        ],
        "summary": "Create a chat completion.",
        "description": "Proxies the request to the configured backend and returns a response\nenriched with privacy and verification metadata.",
        "operationId": "chat_completions",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatCompletionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chat completion response with privacy and verification metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EidolonsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/models": {
      "get": {
        "tags": [
          "Public"
        ],
        "summary": "List available models.",
        "operationId": "list_models",
        "responses": {
          "200": {
            "description": "List of available models",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/webhooks/stripe": {
      "post": {
        "tags": [
          "Linked"
        ],
        "summary": "POST /v1/webhooks/stripe",
        "operationId": "stripe_webhook",
        "responses": {
          "200": {
            "description": "Event received"
          },
          "400": {
            "description": "Signature verification failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Webhook secret not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AssistantMessage": {
        "type": "object",
        "description": "An assistant message in a response.",
        "required": [
          "role"
        ],
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "AuthorizationInfo": {
        "type": "object",
        "description": "Authorization information.",
        "required": [
          "method",
          "linkable"
        ],
        "properties": {
          "linkable": {
            "type": "boolean",
            "description": "Whether this authorization method allows linking across requests."
          },
          "method": {
            "$ref": "#/components/schemas/AuthMethod",
            "description": "The authorization method used."
          }
        }
      },
...
      "BalancePool": {
        "type": "object",
        "required": [
          "amount",
          "source"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "source": {
            "type": "string"
          }
        }
      },
      "BalancesResponse": {
        "type": "object",
        "required": [
          "available",
          "pools"
        ],
        "properties": {
          "available": {
            "type": "integer",
            "format": "int64"
          },
          "pools": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BalancePool"
            }
          }
        }
      },
...
      "ChatCompletionRequest": {
        "type": "object",
        "description": "A chat completion request.",
        "required": [
          "model",
          "messages"
        ],
        "properties": {
          "max_tokens": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The maximum number of tokens to generate.",
            "minimum": 0
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "description": "A list of messages comprising the conversation."
          },
          "model": {
            "type": "string",
            "description": "ID of the model to use."
          },
          "stop": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/StopSequence",
                "description": "Up to 4 sequences where the API will stop generating."
              }
            ]
          },
          "stream": {
            "type": "boolean",
            "description": "Whether to stream partial responses."
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Sampling temperature between 0 and 2."
          },
          "top_p": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Nucleus sampling parameter."
          }
        },
        "additionalProperties": false
      },
...
      "Choice": {
        "type": "object",
        "description": "A completion choice.",
        "required": [
          "index",
          "message"
        ],
        "properties": {
          "finish_reason": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FinishReason",
                "description": "The reason the model stopped generating."
              }
            ]
          },
          "index": {
            "type": "integer",
            "format": "int32",
            "description": "The index of this choice.",
            "minimum": 0
          },
          "message": {
            "$ref": "#/components/schemas/AssistantMessage",
            "description": "The generated message."
          }
        }
      },
...
      "EidolonsResponse": {
        "type": "object",
        "description": "An Eidolons chat completion response.",
        "required": [
          "id",
          "object",
          "created",
          "model",
          "choices",
          "privacy",
          "verification"
        ],
        "properties": {
          "choices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Choice"
            },
            "description": "Completion choices."
          },
          "created": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of when the completion was created.",
            "minimum": 0
          },
          "id": {
            "type": "string",
            "description": "Unique identifier for the completion."
          },
          "model": {
            "type": "string",
            "description": "The model that produced the completion."
          },
          "object": {
            "type": "string",
            "description": "The object type."
          },
          "privacy": {
            "$ref": "#/components/schemas/PrivacyMetadata",
            "description": "Privacy metadata describing data exposure and authorization."
          },
          "usage": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Usage",
                "description": "Token usage statistics."
              }
            ]
          },
          "verification": {
            "$ref": "#/components/schemas/VerificationMetadata",
            "description": "Verification metadata with cryptographic attestations."
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "message",
          "type"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "An error response in OpenAI format.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "FinishReason": {
        "type": "string",
        "description": "The reason the model stopped generating.",
        "enum": [
          "stop",
          "length",
          "content_filter"
        ]
      },
...
      "LedgerEntry": {
        "type": "object",
        "required": [
          "id",
          "delta",
          "reason",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "delta": {
            "type": "integer",
            "format": "int64"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string"
          },
          "token_credits": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "token_epoch": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "LedgerResponse": {
        "type": "object",
        "required": [
          "data",
          "has_more"
        ],
        "properties": {
          "cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LedgerEntry"
            }
          },
          "has_more": {
            "type": "boolean"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "description": "A message in the conversation.",
        "required": [
          "role",
          "content"
        ],
        "properties": {
          "content": {
            "$ref": "#/components/schemas/MessageContent",
            "description": "The content of the message."
          },
          "name": {
            "type": [
              "string",
              "null"
            ],
            "description": "An optional name for the participant."
          },
          "role": {
            "$ref": "#/components/schemas/Role",
            "description": "The role of the message author."
          }
        },
        "additionalProperties": false
      },
      "MessageContent": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentPart"
            }
          }
        ],
        "description": "Message content can be a simple string or array of content parts (for multimodal)."
      },
      "Model": {
        "type": "object",
        "description": "A model descriptor.",
        "required": [
          "id",
          "name",
          "created",
          "description",
          "context_length"
        ],
        "properties": {
          "context_length": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum context window size in tokens.",
            "minimum": 0
          },
          "created": {
            "type": "integer",
            "format": "int64",
            "description": "Unix timestamp of when the model was created.",
            "minimum": 0
          },
          "description": {
            "type": "string",
            "description": "Short description of the model's capabilities."
          },
          "id": {
            "type": "string",
            "description": "The model identifier (e.g. \"openai/gpt-4o\")."
          },
          "name": {
            "type": "string",
            "description": "Human-readable display name."
          }
        }
      },
      "ModelsResponse": {
        "type": "object",
        "description": "A list of available models.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Model"
            },
            "description": "The list of models."
          }
        }
      },
...
      "PrivacyMetadata": {
        "type": "object",
        "description": "Privacy metadata describing how the request was authorized, what data is\nvisible to which parties, and transport security properties.",
        "required": [
          "authorization",
          "data_exposure",
          "transport"
        ],
        "properties": {
          "authorization": {
            "$ref": "#/components/schemas/AuthorizationInfo",
            "description": "Authorization details."
          },
          "data_exposure": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataExposure"
            },
            "description": "Which parties can see which data, and under what guarantees."
          },
          "transport": {
            "$ref": "#/components/schemas/TransportInfo",
            "description": "Transport security between hops."
          }
        }
      },
...
      "Role": {
        "type": "string",
        "description": "The role of a message author.",
        "enum": [
          "system",
          "user",
          "assistant"
        ]
      },
      "StopSequence": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        ],
        "description": "Stop sequence can be a single string or array of strings."
      },
...
      "TransportInfo": {
        "type": "object",
        "description": "Transport security information.",
        "required": [
          "client_to_proxy",
          "proxy_to_backend"
        ],
        "properties": {
          "client_to_proxy": {
            "type": "string",
            "description": "Transport from client to this proxy."
          },
          "proxy_to_backend": {
            "type": "string",
            "description": "Transport from this proxy to the backend."
          }
        }
      },
      "Usage": {
        "type": "object",
        "description": "Token usage statistics.",
        "required": [
          "prompt_tokens",
          "completion_tokens",
          "total_tokens"
        ],
        "properties": {
          "completion_tokens": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "prompt_tokens": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "VerificationMetadata": {
        "type": "object",
        "description": "Verification metadata with cryptographic attestations for the proxy and backend.",
        "properties": {
          "backend": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BackendAttestation",
                "description": "Backend attestation (only for TEE-backed models like phala/*)."
              }
            ]
          },
          "proxy": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProxyAttestation",
                "description": "Proxy attestation (None until the proxy runs in a TEE)."
              }
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic",
        "description": "Account credentials. Username is the account UUID, password is the secret returned at account creation."
      }
    }
  },
...
}